Move Quiz to QuizView
