+++ src/components/layout/Quiz/Flashcards.tsx
@@ -5,6 +5,8 @@ import QuizHeader from "./QuizHeader";
 import { TermProps } from "./Term";
 
 export default function Flashcards() {
+  // TODO: Nie wiem co tutaj zrobic
+
   const [currTerm, setCurrTerm] = useState<number>(0);
 
   const terms: TermProps[] = [
