+++ src/components/layout/Sections/TileGrid/QuizTile.tsx
@@ -1,4 +1,5 @@
 import { QuizResponse } from "@/api/types/quiz";
+import { Link } from "react-router-dom";
 import img from "../../../../assets/images/office.jpg";
 
 type QuizTileProps = {
@@ -9,17 +10,17 @@ export default function QuizTile({ quiz }: QuizTileProps) {
   const url = `/${quiz.userName}/${quiz.slug}`;
   return (
     <div className="bg-secondary rounded-md w-full">
-      <a href={url} className="relative">
+      <Link to={url} className="relative">
         <img src={img} className="w-full rounded-t-md" />
         <div className="text-white text-sm font-bold absolute top-4 right-4 px-4 py-2 bg-lightBlue rounded-3xl">
           Category/QS
         </div>
-      </a>
+      </Link>
       <div className="p-6">
-        <a href={url} className="relative">
+        <Link to={url} className="relative">
           <h3 className="text-white text-2xl font-bold mb-3">{quiz.name}</h3>
           <p className="text-textPrimary font-medium">{quiz.description}</p>
-        </a>
+        </Link>
         <span className="h-[1px] bg-white bg-opacity-10 block my-6"></span>
         <div className="flex items-center gap-4">
           <img src={img} className="w-10 h-10 rounded-3xl" />
